requirements.txt
langchain_openai #For using OpenAI API
langchain_core #For embedding chatprompt template
python-dotenv
streamlit #For UI
langchain_community #For calling open source llms
langserve #For deployments, creating api routes
fastapi 
uvicorn 
sse_starlette
bs4
pypdf
chromadb
faiss-cpu
